@@ -29,7 +29,7 @@ import com.gufli.kingdomcraft.common.ebean.beans.query.QBUser;
 import io.ebean.annotation.ConstraintMode;
 import io.ebean.annotation.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
