@@ -3,7 +3,7 @@ package com.gufli.kingdomcraft.common.ebean.beans;
 import io.ebean.Model;
 import io.ebean.annotation.DbName;
 
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.MappedSuperclass;
 
 @DbName("kingdomcraft")
 @MappedSuperclass
